updated posts components
