+++ src/components/Post/LatestPost.js
@@ -25,7 +25,7 @@ const LatestPost = () => {
             </h4>
             <div className="flex">
               <p className="w-96 truncate">{post.text}</p>
-              <NavLink to="/post"><button className="text-green-500" type="button" onClick={() => setPostId(post.post_id)}>more</button></NavLink>
+              <NavLink to="/postdetails"><button className="text-green-500" type="button" onClick={() => setPostId(post.post_id)}>more</button></NavLink>
             </div>
             <LikeCommentCount id={post.post_id} />
             <div>
